+++ src/main/datetime.c
@@ -306,12 +306,12 @@ static double mktime00(struct tm *tm)
     if (year0 > 3000)
     {
         excess = (int)(year0 / 2000) - 1;
-        year0 -= (int)(excess * 2000);
+        year0 -= excess * 2000;
     }
     else if (year0 < 0)
     {
         excess = -1 - (int)(-year0 / 2000);
-        year0 -= (int)(excess * 2000);
+        year0 -= excess * 2000;
     }
 
     for (i = 0; i < tm->tm_mon; i++)
