+++ src/main/mkdtemp.c
@@ -114,7 +114,7 @@ static int gen_tempname(char *tmpl)
     int count, fd = -1;
     int save_errno = errno;
 
-    len = strlen(tmpl);
+    len = (int)strlen(tmpl);
     if (len < 6 || strcmp(&tmpl[len - 6], "XXXXXX"))
     {
         __set_errno(EINVAL);
